consumerPython/requirements.txt
kafka_python==1.4.3
# kafka==1.3.5
